+++ src/view/point/favorite-button.js
@@ -1,0 +1,32 @@
+import {createElement} from '../render';
+
+export default class FavoriteButton {
+  #TEMPLATE =
+    `<button class="event__favorite-btn" type="button">
+    <span class="visually-hidden">Add to favorite</span>
+    <svg class="event__favorite-icon" width="28" height="28" viewBox="0 0 28 28">
+      <path
+        d="M14 21l-8.22899 4.3262 1.57159-9.1631L.685209 9.67376 9.8855 8.33688 14 0l4.1145 8.33688 9.2003 1.33688-6.6574 6.48934 1.5716 9.1631L14 21z"/>
+    </svg>
+  </button>`;
+
+  #element;
+
+  constructor(active) {
+    this.#element = createElement(this.#TEMPLATE);
+    if (active) {
+      this.#element.classList.add('event__favorite-btn--active');
+    }
+  }
+
+  getElement = () => this.#element;
+
+  set state(active) {
+    if (active) {
+      this.#element.classList.add('event__favorite-btn--active');
+    } else {
+      this.#element.classList.remove('event__favorite-btn--active');
+    }
+    this.#element.replaceWith(this.#element);
+  }
+}
