+++ src/view/point/route-point.js
@@ -2,68 +2,8 @@ import dayjs from 'dayjs';
 import {createElement, createElementSan, RenderPosition} from '../render';
 import {Icons} from '../../settings.js';
 import RollupButton from './rollup-btn.js';
-
-class FavoriteButton {
-  #TEMPL =
-    `<button class="event__favorite-btn" type="button">
-    <span class="visually-hidden">Add to favorite</span>
-    <svg class="event__favorite-icon" width="28" height="28" viewBox="0 0 28 28">
-      <path
-        d="M14 21l-8.22899 4.3262 1.57159-9.1631L.685209 9.67376 9.8855 8.33688 14 0l4.1145 8.33688 9.2003 1.33688-6.6574 6.48934 1.5716 9.1631L14 21z"/>
-    </svg>
-  </button>`;
-
-  #element;
-
-  constructor(active) {
-    this.#element = createElement(this.#TEMPL);
-    if (active) {
-      this.#element.classList.add('event__favorite-btn--active');
-    }
-  }
-
-  getElement = () => this.#element;
-
-  set state(active) {
-    if (active) {
-      this.#element.classList.add('event__favorite-btn--active');
-    } else {
-      this.#element.classList.remove('event__favorite-btn--active');
-    }
-    this.#element.replaceWith(this.#element);
-  }
-}
-
-class Offer {
-  #TEMPL;
-  #element;
-
-  constructor(offer) {
-    this.#TEMPL =
-      `<li class="event__offer">
-      <span class="event__offer-title">${offer.title}</span>
-      &plus;&euro;&nbsp;
-      <span class="event__offer-price">${offer.price}</span>
-    </li>`;
-    this.#element = createElementSan(this.#TEMPL);
-  }
-
-  getElement = () => this.#element;
-}
-
-class listOffers {
-  #TEMPL = '<ul class="event__selected-offers"></ul>';
-  #element;
-
-  constructor(offers) {
-    this.#element = createElement(this.#TEMPL);
-    offers.forEach((offer) => {
-      this.#element.append(new Offer(offer).getElement());
-    });
-  }
-
-  getElement = () => this.#element;
-}
+import FavoriteButton from './favorite-button';
+import ListOffers from './list-offers';
 
 export default class RoutePoint {
   #TEMPLATE;
@@ -139,7 +79,7 @@ export default class RoutePoint {
     );
 
     if (offers) {
-      this.#offers = new listOffers(offers).getElement();
+      this.#offers = new ListOffers(offers).getElement();
       this.#element.append(this.#offers);
     } else {
       this.#offers = false;
@@ -174,7 +114,7 @@ export default class RoutePoint {
           this.#offers.remove();
         }
         if (offers) {
-          this.#offers = new listOffers(offers).getElement();
+          this.#offers = new ListOffers(offers).getElement();
           this.#favoriteButton.getElement().insertAdjacentElement(RenderPosition.BEFOREBEGIN, this.#offers);
         } else {
           this.#offers = false;
@@ -217,5 +157,4 @@ export default class RoutePoint {
       (h > 0 ? `${h.toString(10).padStart(2, '0')}H ` : '')
     }${m.toString().padStart(2, '0')}M`;
   };
-
 }
